organize for production

+++ src/database/models.py
@@ -26,8 +26,9 @@ db_drop_and_create_all()
     !!NOTE you can change the database_filename variable to have multiple verisons of a database
 '''
 def db_drop_and_create_all():
-    db.drop_all()
-    db.create_all()
+    if not os.path.isfile(os.path.join(project_dir, database_filename)):
+        db.drop_all()
+        db.create_all()
 
 '''
 Drink
